Добавлена фабрика получения запросов Определение пользователя и ид чата вынесено в Bot
