+++ model/src/main/java/com/vol/chatbot/dao/MessageDao.java
@@ -1,7 +1,7 @@
 package com.vol.chatbot.dao;
 
-import com.vol.chatbot.model.impl.Message;
-import com.vol.chatbot.model.impl.User;
+import com.vol.chatbot.model.Message;
+import com.vol.chatbot.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Set;
