+++ model/src/main/java/com/vol/chatbot/dao/PropertiesDao.java
@@ -1,7 +1,7 @@
 package com.vol.chatbot.dao;
 
-import com.vol.chatbot.model.impl.Properties;
-import com.vol.chatbot.model.impl.Property;
+import com.vol.chatbot.model.Properties;
+import com.vol.chatbot.model.Property;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
