+++ model/src/main/java/com/vol/chatbot/dao/ScenarioDao.java
@@ -1,6 +1,6 @@
 package com.vol.chatbot.dao;
 
-import com.vol.chatbot.model.impl.Scenario;
+import com.vol.chatbot.model.Scenario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ScenarioDao extends JpaRepository<Scenario, Long> {
