+++ model/src/main/java/com/vol/chatbot/dao/ScenarioStepDao.java
@@ -1,7 +1,7 @@
 package com.vol.chatbot.dao;
 
-import com.vol.chatbot.model.impl.Scenario;
-import com.vol.chatbot.model.impl.ScenarioStep;
+import com.vol.chatbot.model.Scenario;
+import com.vol.chatbot.model.ScenarioStep;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ScenarioStepDao extends JpaRepository<ScenarioStep, Long> {
