+++ model/src/main/java/com/vol/chatbot/dao/UserInfoDao.java
@@ -1,6 +1,6 @@
 package com.vol.chatbot.dao;
 
-import com.vol.chatbot.model.impl.UserInfo;
+import com.vol.chatbot.model.UserInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserInfoDao extends JpaRepository<UserInfo, Long> {
