+++ model/src/main/java/com/vol/chatbot/model/Answer.java
@@ -1,16 +1,48 @@
 package com.vol.chatbot.model;
 
-import com.vol.chatbot.model.impl.User;
+import javax.persistence.*;
 
-public interface Answer {
-    String getAnswer();
+@Entity
+@Table(name = "bot_answer")
+public class Answer{
 
-    void setAnswer(String answer);
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id", unique = true, nullable = false)
+    private Long id;
 
-    User getUser();
+    @Column(name = "answer", unique = false, nullable = true)
+    private String answer;
 
-    void setUser(User user);
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
 
-    Question getQuestion();
+    @ManyToOne
+    @JoinColumn(name = "question_id")
+    private Question question;
 
+    public String getAnswer() {
+        return answer;
+    }
+
+    public void setAnswer(String answer) {
+        this.answer = answer;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+    }
+
+    public Question getQuestion() {
+        return question;
+    }
+
+    public void setQuestion(Question question) {
+        this.question = question;
+    }
 }
