+++ model/src/main/java/com/vol/chatbot/model/QuestionWeight.java
@@ -1,4 +1,4 @@
-package com.vol.chatbot.model.impl;
+package com.vol.chatbot.model;
 
 public enum QuestionWeight {
     DIFFICULT,
