+++ model/src/main/java/com/vol/chatbot/model/User.java
@@ -1,4 +1,4 @@
-package com.vol.chatbot.model.impl;
+package com.vol.chatbot.model;
 
 
 import javax.persistence.*;
@@ -31,12 +31,14 @@ public class User {
     @Column(name = "datecreate")
     private Date datecreate;
 
+    //TODO пока закоментровал так как не понятно как быть с идентификацией польхователем
+ /*
     @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.EAGER, mappedBy = "user")
     private UserInfo userInfo;
 
     @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.EAGER, mappedBy = "user")
     private Scenario scenario;
-
+*/
     public User() {
         // Do nothing
     }
@@ -97,22 +99,22 @@ public class User {
         this.datecreate = datecreate;
     }
 
-    public UserInfo getUserInfo() {
-        return userInfo;
-    }
+    //TODO пока закоментровал так как не понятно как быть с идентификацией польхователем
+        public UserInfo getUserInfo() {
+            throw new UnsupportedOperationException();
+        }
 
-    public void setUserInfo(UserInfo userInfo) {
-        this.userInfo = userInfo;
-    }
+        public void setUserInfo(UserInfo userInfo) {
+            throw new UnsupportedOperationException();
+        }
 
-    public Scenario getScenario() {
-        return scenario;
-    }
+        public Scenario getScenario() {
+            throw new UnsupportedOperationException();
+        }
 
-    public void setScenario(Scenario scenario) {
-        this.scenario = scenario;
-    }
-
+        public void setScenario(Scenario scenario) {
+            throw new UnsupportedOperationException();
+        }
     @Override
     public String toString() {
         return "User{" +
@@ -123,8 +125,9 @@ public class User {
             ", userName='" + userName + '\'' +
             ", bot=" + bot +
             ", datecreate=" + datecreate +
-            ", userInfo=" + userInfo +
-            ", scenario=" + scenario +
+            //TODO пока закоментровал так как не понятно как быть с идентификацией польхователем
+//            ", userInfo=" + userInfo +
+//            ", scenario=" + scenario +
             '}';
     }
