+++ model/src/main/resources/initProperties.sql
@@ -1,2 +1,3 @@
 INSERT INTO public.bot_properties ( property_name, property_value) VALUES ( 'SuspendMode', 'false');
-INSERT INTO public.bot_properties ( property_name, property_value) VALUES ( 'UseDay', '1');
+INSERT INTO public.bot_properties ( property_name, property_value) VALUES ( 'useDay', '1');
+INSERT INTO public.bot_properties ( property_name, property_value) VALUES ( 'timeQuestionRefresh', '60');
