+++ model/src/test/java/com/vol/chatbot/model/MessageTest.java
@@ -1,7 +1,6 @@
 package com.vol.chatbot.model;
 
 import com.vol.chatbot.PojoTestUtils;
-import com.vol.chatbot.model.impl.Message;
 import org.junit.Test;
 
 import static org.junit.Assert.assertNotNull;
