+++ module/src/main/java/com/vol/chatbot/bot/BotService.java
@@ -1,6 +1,6 @@
 package com.vol.chatbot.bot;
 
-import com.vol.chatbot.model.impl.User;
+import com.vol.chatbot.model.User;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
