+++ module/src/main/java/com/vol/chatbot/knowledge/QuestionFactory.java
@@ -5,7 +5,7 @@ import com.vol.chatbot.dao.AnswerDao;
 import com.vol.chatbot.dao.QuestionDao;
 import com.vol.chatbot.model.Answer;
 import com.vol.chatbot.model.Question;
-import com.vol.chatbot.model.impl.User;
+import com.vol.chatbot.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,19 +17,18 @@ public class QuestionFactory {
 
     private AnswerDao answerDao;
     private QuestionDao questionDao;
+    private List<Question> questions;
 
     @Autowired
     public QuestionFactory(AnswerDao answerDao, QuestionDao questionDao) {
         this.answerDao = answerDao;
         this.questionDao = questionDao;
+        this.questions = questionDao.findAll();
     }
-
 
     public Question getQuestion(User user) {
         List<Answer> answers = answerDao.findAllByUser(user);
-        List<Question> questions = questionDao.findAll();
         return questions.get(Utils.getRandomInt(questions.size()));
     }
-
 
 }
