+++ module/src/main/java/com/vol/chatbot/services/ScenarioService.java
@@ -2,8 +2,8 @@ package com.vol.chatbot.services;
 
 import com.vol.chatbot.dao.ScenarioDao;
 import com.vol.chatbot.dao.ScenarioStepDao;
-import com.vol.chatbot.model.impl.Scenario;
-import com.vol.chatbot.model.impl.ScenarioStep;
+import com.vol.chatbot.model.Scenario;
+import com.vol.chatbot.model.ScenarioStep;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
