+++ module/src/main/java/com/vol/chatbot/sql/HibernateExporter.java
@@ -98,7 +98,6 @@ public class HibernateExporter {
         String filename = directory + "ddl_" + dialect.toLowerCase() + ".sql";
         LOGGER.info("Output file name {}", filename);
         export.setOutputFile(filename);
-        export.setImportFiles(filename);
         export.setFormat(true);
 
         //can change the output here
