+++ module/src/main/java/com/vol/chatbot/steps/AskSurname.java
@@ -1,8 +1,8 @@
 package com.vol.chatbot.steps;
 
-import com.vol.chatbot.model.impl.Message;
-import com.vol.chatbot.model.impl.User;
-import com.vol.chatbot.model.impl.UserInfo;
+import com.vol.chatbot.model.Message;
+import com.vol.chatbot.model.User;
+import com.vol.chatbot.model.UserInfo;
 import com.vol.chatbot.services.ScenarioService;
 import com.vol.chatbot.services.UserInfoService;
 import org.springframework.beans.factory.annotation.Autowired;
