+++ module/src/main/java/com/vol/chatbot/steps/StepExecutor.java
@@ -1,7 +1,7 @@
 package com.vol.chatbot.steps;
 
-import com.vol.chatbot.model.impl.Message;
-import com.vol.chatbot.model.impl.User;
+import com.vol.chatbot.model.Message;
+import com.vol.chatbot.model.User;
 
 public interface StepExecutor {
