+++ module/src/main/resources/application.properties
@@ -1,4 +1,4 @@
 spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 queues.outgoing.tps=20.0
-properties.refreshtps=0.5
+properties.refreshtps=0.01
