+++ module/src/main/resources/database.properties
@@ -1,4 +1,4 @@
 jdbc.driverClassName=org.postgresql.Driver
 jdbc.url=jdbc:postgresql://notes.cfcijwxje5wr.us-west-2.rds.amazonaws.com:5432/notes
-jdbc.username=test_user
-jdbc.password=test_user
+jdbc.username=botadmin
+jdbc.password=adminbot
